+++ Bezhkov/Bulk-Thumbnail-Creator/App.config
