+++ Bezhkov/Bulk-Thumbnail-Creator/Choise.cs
