--- Bezhkov/Bulk-Thumbnail-Creator/ConsoleApp14.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31005.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp14", "ConsoleApp14\ConsoleApp14.csproj", "{831A3BF9-01D1-4804-A18A-E9E9566A0B2A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{831A3BF9-01D1-4804-A18A-E9E9566A0B2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{831A3BF9-01D1-4804-A18A-E9E9566A0B2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{831A3BF9-01D1-4804-A18A-E9E9566A0B2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{831A3BF9-01D1-4804-A18A-E9E9566A0B2A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {58682007-CA4E-44E3-8595-FB0A2C705B6D}
-	EndGlobalSection
-EndGlobal
