--- Bezhkov/Bulk-Thumbnail-Creator/ConsoleApp14/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Drawing.Common" version="5.0.1" targetFramework="net472" />
-</packages>
