+++ Bezhkov/Bulk-Thumbnail-Creator/ManagerProgramm.cs
@@ -38,8 +38,10 @@ namespace ConsoleApp14
         }
 
         //Default constructor
-        public ManagerProgramm() 
-        { }
+        public ManagerProgramm()
+        {
+
+        }
 
         //Homework information
         internal void Info()
